# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tech task forum",
        "description": "Api documentation Tech task forum",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "OpenApi host"
        }
    ],
    "paths": {
...
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User login.",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login completed successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AuthUserRequest": {
                "title": "User registration request.",
                "description": "Class AuthUserRequest.",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "description": "Email.",
                        "type": "string",
                        "example": "test@gmail.com"
                    },
                    "password": {
                        "description": "Password.",
                        "type": "string",
                        "example": "12345678"
                    }
                },
                "type": "object"
            },
...
            "AuthResource": {
                "title": "Auth resourse",
                "description": "Class AuthResource.",
                "properties": {
                    "token": {
                        "title": "Token",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Authorization by login and password with obtaining an authorization token.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}